FIRST WORKING INSTANCE OF GRPC WEB CLIENT
Added in the stupid http and https requirements to package.json which didn't even fix the grpc-web-client issue, so I had to make 2 empty files so node can resolve the module that it will never use.

Added in test button to show that unary calls are working to the back end.