Added a default empty lambda for the onEnd parameter in the invoke function of the rpc service. This will allow the user not to provide an onEnd function if it isn't needed, which I suspect it often won't be.
Added in support for the new RPC in the CreatureService, GenerateCreaturesRpc. Still have some tweaking to do with type conversions to get it all just right. All of this "toObject" and "X_Message" as opposed to just "X" has me a bit confused. Will iron these details out going forward. For now a todo will suffice.

Working on converting the creature model. Added support to provide a CreatureMessage to the constructor which will deserialize the raw data into a valid Creature Model (with all the default values and associated functions, instead of just speed stamina health greed name and generation).