With the help of Chau Tran (https://github.com/nartc), we managed to figure out a way to do the gridification in a way to doesn't require doing "this.creatures.concat()". Granted that was only 1 line and this solution is about 40 lines, however this way is more the way it "should" be done, as far as I can tell.
Started working on the simulateCreatures RPC, but got distracted in the middle.

Removed the gridify pipe, as it is no longer needed. Rest in peace, gridify, it was a good idea.

Renamed `message` in each of hte function in the rpc service to `method`, which is a more accurate description of what that is actually for.