Big Changes. Converted all current RPCs to unary because thats the only one that really makes sense at this point.
Added a change to the unary call in the rpc-service which converts the return type from "UnaryOutput" to instead be a direct ProtobufMessage where the parameters following it are the other objects you can grab from the UnaryOutput. This change was made to keep the unary call more in line with the other two types of calls.

Lots of changes to each of the four RPCs to make them more standardized and reflective of each other.

Added and removed some things from the creature model.

At this point all unary versions of generation and simulation appear to be working as intended.

Undid some stupid html formatting done by IntelliJ's auto-wrapping of long lines.