Changed the colors of some buttons.
Focused in on ways to improve the serialization and deserialization of a generated gRPC message class into, what I am now calling, a PB-Class. The Creature class is considered a PB-Class. This is because it effectively wraps a gRPC message into a richer class that contains more functionality. This lead to the creation of a new interface called "pb-serializable" which enforces that any class that implements it (PB-Classes) have both a way to convert raw gRPC messages into themselves (type promotion, "sateFromMessage") and a way to convert themselves into the original gRPC message again (type demotion, "toMessage").

Found a way to tie in the PB-Class to its respective Message. This is done by implementing the Message.AsObject type (weird right, but you can implement types apparently) in the declaration of the PB-Class. This enables intellisense related to both the PB-Class and base class, as well as stricter type checking during compilation, and feels all around a lot better.

Went back through and created an official PBClass model. This is to be extended by any class which will be used for enriching gRPC responses. It gaurantees that the sateFromMessage call happens as soon as the object is instantiated. Apologies for these insane generics, I believe they are correct, but they do feel pretty overkill.