Implemented support for population messages.
Changed the constructor for all PBClasses to include a default value for the parameter (which would just be an empty M, whatever M happens to be.) I would like to have included this behavior in the PBClass itself, but it does not appear to be possible at this point.

Implemented the fitness index variable into the creature model.